refactor: UserSearchBar, UserSearchModal
